requirements.txt
pynput==1.7.6
python-socketio==5.9.0
websocket-client==1.6.2
